WIP Dialogue System
Added Dialogue Manager, a NPC BP, changed the Dialogue Box WB and the text of it, added an NPC to the playground.

Note: We will have to change the UI to be called by an animation and not by setting it hidden or visible, because when the talk index is set to -1 again, it starts the dialogue by showing the last array first and automatically passing to the next one.